requirements.txt
# requirements.txt

# ─── Core LangChain stack ───────────────────────────────────────────────────
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-experimental>=0.0.60

# ─── LLM Providers ──────────────────────────────────────────────────────────
langchain-groq>=0.1.0           # Groq cloud inference
langchain-ollama>=0.1.0         # Local Ollama inference (pip install langchain-ollama)
langchain-openai>=0.1.0         # OpenAI / OpenAI-compatible endpoints

# ─── Embeddings & Reranking ─────────────────────────────────────────────────
langchain-huggingface>=0.0.3
sentence-transformers>=2.7.0    # Multilingual embeddings (supports Arabic)
transformers>=4.40.0
torch>=2.2.0

# ─── Vector Store ───────────────────────────────────────────────────────────
chromadb>=0.5.0

# ─── Hybrid Search ──────────────────────────────────────────────────────────
elasticsearch>=8.0.0
langchain-elasticsearch>=0.2.0

# ─── Knowledge Graph ────────────────────────────────────────────────────────
neo4j>=5.0.0

# ─── Document Loading (Arabic-aware) ────────────────────────────────────────
pypdf>=4.0.0                    # PDF extraction
python-docx>=1.0.0              # Arabic .docx support
lxml>=5.0.0

# ─── API Layer ──────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# ─── Data validation & configuration ────────────────────────────────────────
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# ─── Web search tools ───────────────────────────────────────────────────────
duckduckgo-search>=6.0.0
wikipedia>=1.4.0
google-search-results>=2.4.0    # Optional: Serper / Google Search

# ─── Utilities ──────────────────────────────────────────────────────────────
tiktoken>=0.7.0
numexpr>=2.8.0
requests>=2.31.0

# ─── Optional: Advanced Arabic NLP ──────────────────────────────────────────
# camel-tools>=1.5.0            # Uncomment for advanced Arabic morphology/stemming
